Moved 2D NMR plotting class out of CLI
